deploy/custom_resources/custom_authorizer/requirements.txt
certifi==2023.7.22
charset-normalizer==3.1.0
ecdsa==0.18.0
idna==3.4
pyasn1==0.5.0
python-jose==3.3.0
requests==2.31.0
rsa==4.9
six==1.16.0
urllib3==1.26.18